Redone Object
- Object is simpler. Using table as constructors makes everything easier
because [named arguments] are better (easier to understand what derrek
you are later) and also makes it possible to use [optional arguments],
with or without default values. Awesome!
- Redone stuff that depended on it (thankfully not much for now)
- Really really small tweaks
PS. I recommend using 'class declarations' as a way to show all
arguments that need to be set when creating an object of that type
(Timer and Sprite do that)